requirements.txt

google-generativeai
python-dotenv
pydantic
pyyaml
jsonschema
